requirements.txt
fastapi
uvicorn
google-cloud-firestore
pydantic
pydantic_settings
python-dotenv
python-jose